prepare for use of parentDTO

+++ src/main/java/com/jj/Gradebook/service/parent/ParentServiceImpl.java
@@ -2,6 +2,7 @@ package com.jj.Gradebook.service.parent;
 
 import com.jj.Gradebook.dao.ParentRepository;
 import com.jj.Gradebook.entity.Parent;
+import com.jj.Gradebook.exceptions.EntityAlreadyExistException;
 import com.jj.Gradebook.exceptions.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
@@ -39,14 +40,26 @@ public class ParentServiceImpl implements ParentService{
     }
 
     @Override
-    public Parent save(Parent parent) {
+    public Parent save(Parent parent) throws EntityAlreadyExistException {
+        Optional<Parent> existingParent = parentRepository.findParentByUser_Pesel(parent.getUser().getPesel());
+        if (existingParent.isEmpty()){
+            Parent savedParent = parentRepository.save(parent);
+            return savedParent;
+        }
+        else {
+            throw new EntityAlreadyExistException("Parent already exists!");
+        }
 
-
-        return parentRepository.save(parent);
     }
 
     @Override
-    public void deleteById(Long id) {
-        parentRepository.deleteById(id);
+    public void deleteById(Long id) throws EntityNotFoundException {
+        Optional<Parent> existingParent = parentRepository.findById(id);
+        if (existingParent.isPresent()) {
+            parentRepository.deleteById(id);
+        }
+        else {
+            throw new EntityNotFoundException("No parent with id - " + id);
+        }
     }
 }
